Move all payment method list creation into PaymentMethodList
